adding/refactoring the code after 1- PR review
